refactor: split root view supporting types
