+++ dogArea/Views/GlobalViews/BaseView/GuestDataUpgradeResultBanner.swift
@@ -1,0 +1,54 @@
+import SwiftUI
+
+struct GuestDataUpgradeResultBanner: View {
+    let report: GuestDataUpgradeReport
+    let onRetry: () -> Void
+    let onDismiss: () -> Void
+
+    private var validationText: String? {
+        guard let passed = report.validationPassed else { return nil }
+        return passed ? "원격 검증 통과" : "원격 검증 실패: \(report.validationMessage ?? "mismatch")"
+    }
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 6) {
+            Text(report.hasOutstandingWork ? "데이터 이관 진행 중" : "데이터 이관 완료")
+                .font(.appFont(for: .SemiBold, size: 13))
+            Text(
+                "세션 \(report.sessionCount)건 · 포인트 \(report.pointCount)건 · \(report.totalAreaM2.calculatedAreaString)"
+            )
+            .font(.appFont(for: .Light, size: 11))
+            .foregroundStyle(Color.appTextDarkGray)
+            if let validationText {
+                Text(validationText)
+                    .font(.appFont(for: .Light, size: 11))
+                    .foregroundStyle(report.validationPassed == true ? Color.appGreen : Color.appRed)
+            }
+            HStack(spacing: 10) {
+                if report.hasOutstandingWork {
+                    Button("재시도") {
+                        onRetry()
+                    }
+                    .font(.appFont(for: .SemiBold, size: 11))
+                    .padding(.horizontal, 10)
+                    .padding(.vertical, 6)
+                    .background(Color.appYellow)
+                    .cornerRadius(8)
+                }
+                Button("닫기") {
+                    onDismiss()
+                }
+                .font(.appFont(for: .SemiBold, size: 11))
+                .padding(.horizontal, 10)
+                .padding(.vertical, 6)
+                .background(Color.appYellowPale)
+                .cornerRadius(8)
+            }
+        }
+        .padding(.horizontal, 12)
+        .padding(.vertical, 10)
+        .background(Color.white.opacity(0.95))
+        .cornerRadius(12)
+        .padding(.horizontal, 12)
+    }
+}
